refactor(images): change image format from PNG to SVG and update their sizes
feat(app): update sizes of some parent containers
