+++ src/Components/Ui/Verification/VerificationEmail.jsx
@@ -1,6 +1,6 @@
 import React , {useContext}from 'react'
-import logoName from '../../../assets/Images/logoEduaName.png';
-import verificationEmailImg from '../../../assets/Images/verificationEmail.png'
+import logoName from '../../../assets/Images/logoEduaName.svg';
+import verificationEmailImg from '../../../assets/Images/email.svg'
 import { Inputs } from '../Inputs/Inputs';
 import { Buttons } from '../Buttons/Buttons';
 import { useLocation } from 'wouter';
@@ -17,10 +17,11 @@ export const VerificationEmail = () => {
         setLocation(`/verificationPhone`);
      };
   return (
-    <div className='bg-white h-screen relative flex flex-col'>
-        <div className='flex justify-center items-center'>
-         <img src={logoName} alt="logo" className="w-[160px] h-[70px]" />
-        </div>
+    <>
+    <div className='bg-white h-full relative my-2'>
+          <div className='flex justify-center items-center'>
+            <img src={logoName} alt="logo" className="w-[170px] h-[55px]" />
+          </div>
         <div className='flex items-center justify-center font-semibold flex-col'>
             <h1 className='text-[30px]'>Verificación en dos pasos </h1>
             <p className='text-[20px]'> Hemos enviado un codigo para verificar tu direccion de correo. Revisa (correo) e ingresa el codigo aquí: </p>
@@ -35,7 +36,7 @@ export const VerificationEmail = () => {
             <Inputs classP={'w-[70px] h-[70px]'}/>
         </div>
         <div className='flex items-center justify-center'>
-            <img src={verificationEmailImg} alt="Email Verification" className='w-60 h-[210px]'/>
+            <img src={verificationEmailImg} alt="Email Verification" className='w-60 h-[225px]'/>
         </div>
         <div className='flex justify-center items-center'>
             <Buttons buttonEvent={() => (setViewSuccessModal(true))} btnStyle={"mt-5 bg-[#D9D9D9] w-[120px]"} label="Enviar código" /> 
@@ -47,7 +48,8 @@ export const VerificationEmail = () => {
         <div className='flex justify-center items-center mt-1'>
         <button onClick={handleButtonClickOne} className='underline cursor-pointer text-[#1E1E1E] font-semibold'>Prueba de otra manera</button>
         </div> 
-        <ModalSuccess visibility={viewSuccessModal}/>
     </div>
+        <ModalSuccess visibility={viewSuccessModal}/>
+    </>
   )
 }
